+++ resolwe/flow/tests/test_annotations.py
@@ -692,13 +692,13 @@ class AnnotationViewSetsTest(TestCase):
             entity=created_value.entity, field=created_value.field
         )
         expected = AnnotationValueSerializer([updated_value], many=True).data
-        print("Got response", response.data)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertCountEqual(response.data, expected)
         with self.assertRaises(AnnotationValue.DoesNotExist):
-            AnnotationValue.objects.get(
-                entity=self.annotation_value1.entity, field=self.annotation_value1.field
-            )
+            AnnotationValue.objects.filter(
+                entity=self.annotation_value1.entity,
+                field=self.annotation_value1.field,
+            ).remove_delete_markers().get()
         created_value = AnnotationValue.objects.get(
             entity=created_value.entity, field=created_value.field
         )
@@ -1451,7 +1451,7 @@ class AnnotationViewSetsTest(TestCase):
         force_authenticate(request, self.contributor)
         response = viewset(request, pk=self.entity1.pk)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(self.entity1.annotations.count(), 1)
+        self.assertEqual(self.entity1.annotations.remove_delete_markers().count(), 1)
         has_value(self.entity1, self.annotation_field2.pk, 2)
 
         # Re-create deleted annotation value.